refactor: StringBuf so it can be included in functions.h
