+++ src/deluge/gui/menu_item/runtime_feature/devSysexSetting.cpp
@@ -20,6 +20,7 @@
 #include "gui/menu_item/runtime_feature/setting.h"
 #include "gui/ui/sound_editor.h"
 #include "model/settings/runtime_feature_settings.h"
+#include "util/d_string.h"
 #include "util/functions.h"
 #include <algorithm>
 #include <iterator>
